add flexible chackVersion functionallity
