+++ server/utils/auth.go
@@ -1,0 +1,104 @@
+package utils
+
+import (
+	"crypto/aes"
+	"crypto/cipher"
+	"crypto/rand"
+	"encoding/base64"
+	"errors"
+	"net/http"
+	"strings"
+
+	"github.com/gin-gonic/gin"
+	"github.com/spf13/viper"
+	"go.mongodb.org/mongo-driver/bson"
+	"go.mongodb.org/mongo-driver/mongo"
+	"golang.org/x/crypto/bcrypt"
+)
+
+func AuthMiddleware(db *mongo.Database) gin.HandlerFunc {
+	return func(c *gin.Context) {
+		authHeader := c.GetHeader("Authorization")
+		tokenParts := strings.Fields(authHeader)
+		if authHeader == "" {
+			c.AbortWithStatusJSON(http.StatusUnauthorized, gin.H{"error": "missing authorization header"})
+			return
+		} else if len(tokenParts) != 2 {
+			c.AbortWithStatusJSON(http.StatusUnauthorized, gin.H{"error": "invalid token"})
+			return
+		}
+
+		bytes, err := DecryptUserCredentials(tokenParts[1])
+		if err != nil {
+			c.AbortWithStatusJSON(http.StatusUnauthorized, gin.H{"error": "invalid token"})
+			return
+		}
+
+		// extract the username and password from the decrypted bytes
+		pair := strings.SplitN(string(bytes), ":", 2)
+		if len(pair) != 2 {
+			c.AbortWithStatusJSON(http.StatusUnauthorized, gin.H{"error": "invalid token"})
+			return
+		}
+		username := pair[0]
+		password := pair[1]
+
+		// check the user credentials against the admins collection in MongoDB
+		admins := db.Collection("admins")
+		var result bson.M
+		err = admins.FindOne(c.Request.Context(), bson.M{"username": username}).Decode(&result)
+		if err != nil {
+			c.AbortWithStatusJSON(http.StatusUnauthorized, gin.H{"error": "invalid username or password"})
+			return
+		}
+
+		// compare the hashed passwords
+		hashedPassword := result["password"].(string)
+		if err := bcrypt.CompareHashAndPassword([]byte(hashedPassword), []byte(password)); err != nil {
+			c.AbortWithStatusJSON(http.StatusUnauthorized, gin.H{"error": "invalid username or password"})
+			return
+		}
+
+		// set the authenticated user in the request context for later use
+		c.Set("username", username)
+		c.Next()
+	}
+}
+
+func EncryptUserCredentials(data []byte) (string, error) {
+	block, err := aes.NewCipher([]byte(viper.GetViper().GetString("SYSTEM_KEY")))
+	if err != nil {
+		return "", err
+	}
+	gcm, err := cipher.NewGCM(block)
+	if err != nil {
+		return "", err
+	}
+	nonce := make([]byte, gcm.NonceSize())
+	if _, err := rand.Read(nonce); err != nil {
+		return "", err
+	}
+	ciphertext := gcm.Seal(nonce, nonce, data, nil)
+	return base64.URLEncoding.EncodeToString(ciphertext), nil
+}
+
+func DecryptUserCredentials(token string) ([]byte, error) {
+	ciphertext, err := base64.URLEncoding.DecodeString(token)
+	if err != nil {
+		return nil, err
+	}
+	block, err := aes.NewCipher([]byte(viper.GetViper().GetString("SYSTEM_KEY")))
+	if err != nil {
+		return nil, err
+	}
+	gcm, err := cipher.NewGCM(block)
+	if err != nil {
+		return nil, err
+	}
+	if len(ciphertext) < gcm.NonceSize() {
+		return nil, errors.New("invalid ciphertext")
+	}
+	nonce := ciphertext[:gcm.NonceSize()]
+	ciphertext = ciphertext[gcm.NonceSize():]
+	return gcm.Open(nil, nonce, ciphertext, nil)
+}
